Change TimeInSeconds type from int to double.

+++ Models/SortingGameData.cs
@@ -4,6 +4,6 @@
     {
         public int Id { get; set; }
         public string? Username { get; set; }
-        public int TimeInSeconds { get; set; }
+        public double TimeInSeconds { get; set; }
     }
 }
